update seed nodes
Replacing mainnet seed nodes with testnet port with the correct (now-working) correct mainnet port.
Kee wants to replace imaginary.seed due to SSL issues, it was using the IP, I've replaced it with the IP for public.loki.foundation for now. 

Not sure where you configure the android testnet config, so I'm not sure I did this right...

But when we enable SSL, we'll need hosts. And for the dude having DNS issues with our seeds, we will either need to disable SSL verification or keep an http port open for him